cyclotomic_decagonal: implement fast vis computation
Using a maxstep value of 400, a discard probability of 0.5, this
improves performance from 6m17.595s to 2.235s (factor = 168).

The only disadvantage of fastvis is that the 'length' information
of a vertex is lost.

TODO: Try to implement this by storing the reduction exponent
(see vec4i::directL5ToUnique) when processing the vertices.