chiral_radial: optimise planar chair mode
When computing vertices for radial projection we need to ensure that
the patch has circular shape. We currently achieve this by applying
a cutoff once all inflation steps are applied.

Optimise this by discarding tiles already durin the inflation phase.
Tiles that are outside the reference circle don't contribute to
the vertices of the circular patch.